Move Dagger component and module to a dedicated package
