+++ app/src/main/java/com/nutomic/syncthingandroid/di/SyncthingModule.java
@@ -1,8 +1,9 @@
-package com.nutomic.syncthingandroid;
+package com.nutomic.syncthingandroid.di;
 
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
 
+import com.nutomic.syncthingandroid.SyncthingApp;
 import com.nutomic.syncthingandroid.service.NotificationHandler;
 
 import javax.inject.Singleton;
